+++ src/components/CategoryList/CategoryList.scss
