+++ src/components/EditTask/EditTask.scss
@@ -41,6 +41,7 @@
 .EditTask-description_input {
   margin-top: 1rem;
   border: solid thin;
+  border-color: lightskyblue;
   background-color: rgb(235, 235, 235);
   width: 90%;
   height: 100%;
