+++ src/components/ProgressBar/ProgressBar.scss
