+++ src/components/Task/Task.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faPen } from "@fortawesome/free-solid-svg-icons";
-import "../assets/scss/Task.scss";
+import "./Task.scss";
 
 const Task = ({ task, setEditingTaskMode, setChoosenTask }) => {
   const onClickEditTask = () => {
